+++ packages/kit/src/core/config/options.js
@@ -88,7 +88,7 @@ const options = object(
 			// TODO: remove this for the 1.0 release
 			amp: error(
 				(keypath) =>
-					`${keypath} has been removed. See https://kit.svelte.dev/docs/seo#amp for details on how to support AMP`
+					`${keypath} has been removed. See https://kit.svelte.dev/docs/seo#manual-setup-amp for details on how to support AMP`
 			),
 
 			appDir: validate('_app', (input, keypath) => {
@@ -258,22 +258,39 @@ const options = object(
 				// TODO: remove this for the 1.0 release
 				force: validate(undefined, (input, keypath) => {
 					if (typeof input !== 'undefined') {
-						const newSetting = input ? 'continue' : 'fail';
-						const needsSetting = newSetting === 'continue';
+						const new_input = input ? 'warn' : 'fail';
+						const needs_option = new_input === 'warn';
 						throw new Error(
-							`${keypath} has been removed in favor of \`onError\`. In your case, set \`onError\` to "${newSetting}"${
-								needsSetting ? '' : ' (or leave it undefined)'
+							`${keypath} has been removed in favor of \`handleHttpError\`. In your case, set \`handleHttpError\` to "${new_input}"${
+								needs_option ? '' : ' (or leave it undefined)'
 							} to get the same behavior as you would with \`force: ${JSON.stringify(input)}\``
 						);
 					}
 				}),
 
-				onError: validate('fail', (input, keypath) => {
+				handleHttpError: validate('fail', (input, keypath) => {
 					if (typeof input === 'function') return input;
-					if (['continue', 'fail'].includes(input)) return input;
-					throw new Error(
-						`${keypath} should be either a custom function or one of "continue" or "fail"`
-					);
+					if (['fail', 'warn', 'ignore'].includes(input)) return input;
+					throw new Error(`${keypath} should be "fail", "warn", "ignore" or a custom function`);
+				}),
+
+				handleMissingId: validate('fail', (input, keypath) => {
+					if (typeof input === 'function') return input;
+					if (['fail', 'warn', 'ignore'].includes(input)) return input;
+					throw new Error(`${keypath} should be "fail", "warn", "ignore" or a custom function`);
+				}),
+
+				// TODO: remove this for the 1.0 release
+				onError: validate(undefined, (input, keypath) => {
+					if (typeof input !== 'undefined') {
+						let message = `${keypath} has been renamed to \`handleHttpError\``;
+
+						if (input === 'continue') {
+							message += ', and "continue" has been renamed to "warn"';
+						}
+
+						throw new Error(message);
+					}
 				}),
 
 				origin: validate('http://sveltekit-prerender', (input, keypath) => {
@@ -321,7 +338,7 @@ const options = object(
 			// TODO remove this for 1.0
 			ssr: error(
 				(keypath) =>
-					`${keypath} has been removed — use the handle hook instead: https://kit.svelte.dev/docs/hooks#handle`
+					`${keypath} has been removed — use the handle hook instead: https://kit.svelte.dev/docs/hooks#server-hooks-handle`
 			),
 
 			// TODO remove this for 1.0