+++ packages/kit/src/utils/url.spec.js
@@ -2,6 +2,9 @@ import * as assert from 'uvu/assert';
 import { describe } from './unit_test.js';
 import { resolve, normalize_path, make_trackable, disable_search } from './url.js';
 
+// @ts-expect-error define global required in url.js
+globalThis.__SVELTEKIT_BROWSER__ = false;
+
 describe('resolve', (test) => {
 	test('resolves a root-relative path', () => {
 		assert.equal(resolve('/a/b/c', '/x/y/z'), '/x/y/z');
@@ -49,6 +52,10 @@ describe('resolve', (test) => {
 
 	test('handles schemes like tel: and mailto:', () => {
 		assert.equal(resolve('/a/b/c', 'mailto:hello@svelte.dev'), 'mailto:hello@svelte.dev');
+	});
+
+	test('resolves a fragment link', () => {
+		assert.equal(resolve('/a/b/c', '#foo'), '/a/b/c#foo');
 	});
 });